pembatasan akses user

+++ app/Http/Controllers/Admin/ReportsController.php
@@ -194,12 +194,26 @@ class ReportsController extends Controller
     }
 
     public function show(Report $report, User $user){  
-                 
+         
+        $rep = $report->user_id;
+        $user = Auth::user()->id;        
+        $cek = $report->follower->find($user);
+
+        
+
         $follower = Follower::with(['userfoll'])->where('report_id', $report->id)->get();
         return view('admin.report.show', compact('follower', 'report'));
     }
 
     public function edit(Report $report){
+
+        // $rep = $report->user_id;
+        // $user = Auth::user()->id;        
+        // $cek = $report->follower->find($user);
+        
+        // if (is_null($cek) && $user !== $rep) {
+        //    return abort(403, Auth::user()->name.' Mo Apa Ko ! Mending Ko Balik -_-');
+        // }
  
          $reports = Report::get();
          $user = User::orderBy('name')->get();
@@ -356,6 +370,14 @@ class ReportsController extends Controller
     }   
 
     public function delete(Report $report){
+
+        $rep = $report->user_id;
+        $user = Auth::user()->id;        
+        $cek = $report->follower->find($user);
+        
+        if (is_null($cek) && $user !== $rep) {
+           return abort(403, Auth::user()->name.' Mo Apa Ko ! Mending Ko Balik -_-');
+        }
 
         Storage::disk('public')->delete(['lainnya/' . $report->documentation->dokumentasi1]);
         Storage::disk('public')->delete(['lainnya/' . $report->documentation->dokumentasi2]);
